requirements.txt
beautifulsoup4==4.9.1
lxml==4.5.2
PyMySQL==0.10.0
requests==2.24.0
python-dateutil==2.8.1
PyYAML==5.4.1
pytest==6.2.4
pytest-html==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==10.2




